removed hardcoding file for posts. unneeded

+++ src/main/java/com/codeup/uswomenvets/dao/Posts.java
@@ -8,7 +8,7 @@ public interface Posts {
     // get a list of all the posts
     List<Post> all(String q);
     // insert a new post and return the new post's id
-    Long insert(Post post);
+    int insert(Post post);
 
     List <Post> userPost(long userId);
